fix: eliminate signed left-shift UB in ngtcp2_ntohl/ngtcp2_ntohs on Windows

+++ lib/ngtcp2_net.h
@@ -114,9 +114,9 @@ STIN uint16_t ngtcp2_htons(uint16_t hostshort) {
 STIN uint32_t ngtcp2_ntohl(uint32_t netlong) {
   uint32_t res;
   unsigned char *p = (unsigned char *)&netlong;
-  res = (uint32_t)(*p++ << 24);
-  res += (uint32_t)(*p++ << 16);
-  res += (uint32_t)(*p++ << 8);
+  res = ((uint32_t)*p++) << 24;
+  res += ((uint32_t)*p++) << 16;
+  res += ((uint32_t)*p++) << 8;
   res += *p;
   return res;
 }
@@ -124,7 +124,7 @@ STIN uint32_t ngtcp2_ntohl(uint32_t netlong) {
 STIN uint16_t ngtcp2_ntohs(uint16_t netshort) {
   uint16_t res;
   unsigned char *p = (unsigned char *)&netshort;
-  res = (uint16_t)(*p++ << 8);
+  res = (uint16_t)(((uint16_t)*p++) << 8);
   res += *p;
   return res;
 }
